task12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	unsigned long x;
+	unsigned int x;
 
 	x = n;
 	if (n < 0)
